api/requirements.txt
boto3>=1.34.0
python-jose[cryptography]>=3.3.0
stripe>=8.0.0
python-ulid>=2.0.0
bcrypt>=4.0.0
